requirements.txt
transformers==4.50.3
torch==2.6.0
sentence-transformers==2.6.1
pymupdf==1.25.5
nltk==3.9.1
faiss-cpu==1.10.0
gradio==5.23.3
requests==2.32.3
langchain==0.3.23
accelerate>=0.26.0